# Advisor: Run check steps in parallel