--- a/apps/advisor/pkg/app/checks/ifaces.go
+++ b/apps/advisor/pkg/app/checks/ifaces.go
@@ -24,6 +24,6 @@ type Step interface {
 	Title() string
 	// Description returns the description of the step
 	Description() string
-	// Run executes the step and returns a list of errors
-	Run(ctx context.Context, obj *advisorv0alpha1.CheckSpec, items []any) ([]advisorv0alpha1.CheckReportError, error)
+	// Run executes the step over an item
+	Run(ctx context.Context, obj *advisorv0alpha1.CheckSpec, item any) (*advisorv0alpha1.CheckReportError, error)
 }