--- a/apps/advisor/pkg/app/utils_test.go
+++ b/apps/advisor/pkg/app/utils_test.go
@@ -3,6 +3,7 @@ package app
 import (
 	"context"
 	"errors"
+	"fmt"
 	"testing"
 
 	"github.com/grafana/grafana-app-sdk/resource"
@@ -68,11 +69,43 @@ func TestProcessCheck(t *testing.T) {
 	meta.SetCreatedBy("user:1")
 	client := &mockClient{}
 	ctx := context.TODO()
-	check := &mockCheck{}
+	check := &mockCheck{
+		items: []any{"item"},
+	}
 
 	err = processCheck(ctx, client, obj, check)
 	assert.NoError(t, err)
 	assert.Equal(t, "processed", obj.GetAnnotations()[statusAnnotation])
+}
+
+func TestProcessMultipleCheckItems(t *testing.T) {
+	obj := &advisorv0alpha1.Check{}
+	obj.SetAnnotations(map[string]string{})
+	meta, err := utils.MetaAccessor(obj)
+	if err != nil {
+		t.Fatal(err)
+	}
+	meta.SetCreatedBy("user:1")
+	client := &mockClient{}
+	ctx := context.TODO()
+	items := make([]any, 100)
+	for i := range items {
+		if i%2 == 0 {
+			items[i] = fmt.Sprintf("item-%d", i)
+		} else {
+			items[i] = "error"
+		}
+	}
+	check := &mockCheck{
+		items: items,
+	}
+
+	err = processCheck(ctx, client, obj, check)
+	assert.NoError(t, err)
+	assert.Equal(t, "processed", obj.GetAnnotations()[statusAnnotation])
+	r := client.lastValue.(advisorv0alpha1.CheckV0alpha1StatusReport)
+	assert.Equal(t, r.Count, int64(100))
+	assert.Len(t, r.Errors, 50)
 }
 
 func TestProcessCheck_AlreadyProcessed(t *testing.T) {
@@ -98,7 +131,8 @@ func TestProcessCheck_RunError(t *testing.T) {
 	ctx := context.TODO()
 
 	check := &mockCheck{
-		err: errors.New("run error"),
+		items: []any{"item"},
+		err:   errors.New("run error"),
 	}
 
 	err = processCheck(ctx, client, obj, check)
@@ -108,14 +142,17 @@ func TestProcessCheck_RunError(t *testing.T) {
 
 type mockClient struct {
 	resource.Client
+	lastValue any
 }
 
 func (m *mockClient) PatchInto(ctx context.Context, id resource.Identifier, req resource.PatchRequest, opts resource.PatchOptions, obj resource.Object) error {
+	m.lastValue = req.Operations[0].Value
 	return nil
 }
 
 type mockCheck struct {
-	err error
+	err   error
+	items []any
 }
 
 func (m *mockCheck) ID() string {
@@ -123,7 +160,7 @@ func (m *mockCheck) ID() string {
 }
 
 func (m *mockCheck) Items(ctx context.Context) ([]any, error) {
-	return []any{}, nil
+	return m.items, nil
 }
 
 func (m *mockCheck) Steps() []checks.Step {
@@ -136,9 +173,13 @@ type mockStep struct {
 	err error
 }
 
-func (m *mockStep) Run(ctx context.Context, obj *advisorv0alpha1.CheckSpec, items []any) ([]advisorv0alpha1.CheckReportError, error) {
+func (m *mockStep) Run(ctx context.Context, obj *advisorv0alpha1.CheckSpec, items any) (*advisorv0alpha1.CheckReportError, error) {
 	if m.err != nil {
 		return nil, m.err
+	}
+	i := items.(string)
+	if i == "error" {
+		return &advisorv0alpha1.CheckReportError{}, nil
 	}
 	return nil, nil
 }